+++ chemclipse/plugins/org.eclipse.chemclipse.xxd.converter.supplier.chemclipse/src/org/eclipse/chemclipse/wsd/converter/supplier/chemclipse/internal/io/ChromatogramWriter_1100.java
@@ -107,7 +107,7 @@ public class ChromatogramWriter_1100 extends AbstractChromatogramWriter implemen
 			dataOutputStream.writeInt(scanSignalTotal);
 			for(int signal = 0; signal < scanSignalTotal; signal++) {
 				IScanSignalWSD scanSignal = scanWSD.getScanSignal(signal);
-				int wavelength = scanSignal.getWavelength();
+				int wavelength = (int)scanSignal.getWavelength(); // Change to double 1.2.0.0
 				float abundance = scanSignal.getAbundance();
 				dataOutputStream.writeInt(wavelength);
 				dataOutputStream.writeFloat(abundance);
@@ -187,7 +187,7 @@ public class ChromatogramWriter_1100 extends AbstractChromatogramWriter implemen
 			dataOutputStream.writeInt(scanSignalTotal);
 			for(int signal = 0; signal < scanSignalTotal; signal++) {
 				IScanSignalWSD scanSignal = scanWSD.getScanSignal(signal);
-				int wavelength = scanSignal.getWavelength();
+				int wavelength = (int)scanSignal.getWavelength(); // Change to double 1.2.0.0
 				float abundance = scanSignal.getAbundance();
 				dataOutputStream.writeInt(wavelength);
 				dataOutputStream.writeFloat(abundance);
